tests: move cql_value.rs tests into serialization.rs
